UDP Client: Support writing large reports
If a report is larger than 65536 it will be rejected by kernel and PHP engine and never sent. With this, we ensure that the packet is always sent and that it has proper signaling (MSG_EOR or MSG_EOF).

TLDR; Solves the following issue:

```PHP Warning:  socket_sendto(): unable to write to socket [90]: Message too long```